backend/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
sqlalchemy
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
psycopg2-binary
httpx
